Fixed hangar node activation, sped up fueling

+++ Assets/Scripts/DataModels/Ships/Ship.cs
@@ -26,6 +26,6 @@ public class Ship : ScriptableObject
 
     public float GetFuelPercent()
     {
-        return currentFuel / maxFuel;
+        return (float) currentFuel / maxFuel;
     }
 }
